finished journey with poms
