+++ exercises/exercise-api/src/main/java/kvansipto/exercise/dto/BaseDto.java
@@ -1,0 +1,9 @@
+package kvansipto.exercise.dto;
+
+import lombok.Data;
+
+@Data
+public abstract class BaseDto {
+
+  private String id;
+}
