+++ exercises/exercise-api/src/main/java/kvansipto/exercise/dto/ExerciseDto.java
@@ -1,0 +1,15 @@
+package kvansipto.exercise.dto;
+
+import java.io.Serializable;
+import lombok.Data;
+import lombok.Value;
+
+@Data
+public class ExerciseDto extends BaseDto {
+
+//  long id;
+  String name;
+  String description;
+  String videoUrl;
+  String imageUrl;
+}
