+++ exercises/exercise-api/src/main/java/kvansipto/exercise/resources/ExerciseApi.java
@@ -1,0 +1,9 @@
+package kvansipto.exercise.resources;
+
+import java.util.List;
+import kvansipto.exercise.dto.ExerciseResultDto;
+
+public interface ExerciseApi {
+
+  List<ExerciseResultDto> getResults();
+}
