+++ exercises/exercise-microservice/src/main/java/microservice/entity/MuscleGroup.java
@@ -1,0 +1,19 @@
+package microservice.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum MuscleGroup {
+  CHEST("Грудные"),
+  BACK("Спина"),
+  LEGS("Ноги"),
+  ARMS("Руки"),
+  SHOULDERS("Плечи"),
+  CORE("Пресс");
+
+  private final String name;
+
+  MuscleGroup(String name) {
+    this.name = name;
+  }
+}
