+++ exercises/exercise-microservice/src/main/java/microservice/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package microservice.repository;
+
+import io.project.kvansiptobot.model.User;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<User, Long> {
+
+}
