--- src/main/java/io/project/kvansiptobot/KvansiptoBotApplication.java
@@ -1,13 +1,0 @@
-package io.project.kvansiptobot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KvansiptoBotApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KvansiptoBotApplication.class, args);
-	}
-
-}
