--- src/main/java/io/project/kvansiptobot/model/User.java
@@ -1,53 +1,0 @@
-package io.project.kvansiptobot.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import java.sql.Timestamp;
-import java.util.Objects;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import org.hibernate.proxy.HibernateProxy;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@Entity(name = "users")
-public class User {
-
-  @Id
-  private Long chatId;
-
-  private String userName;
-  private String firstName;
-  private String lastName;
-  private Timestamp registeredAt;
-
-  @Override
-  public final boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null) {
-      return false;
-    }
-    Class<?> oEffectiveClass = o instanceof HibernateProxy
-        ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
-    Class<?> thisEffectiveClass = this instanceof HibernateProxy
-        ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass()
-        : this.getClass();
-    if (thisEffectiveClass != oEffectiveClass) {
-      return false;
-    }
-    User user = (User) o;
-    return getChatId() != null && Objects.equals(getChatId(), user.getChatId());
-  }
-
-  @Override
-  public final int hashCode() {
-    return this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass()
-        .hashCode() : getClass().hashCode();
-  }
-}
