--- src/main/java/io/project/kvansiptobot/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package io.project.kvansiptobot.repository;
-
-import io.project.kvansiptobot.model.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-
-}
