--- src/test/java/io/project/kvansiptobot/KvansiptoBotApplicationTests.java
@@ -1,13 +1,0 @@
-package io.project.kvansiptobot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KvansiptoBotApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
