+++ telegram/src/main/java/kvansipto/telegram/microservice/config/BotConfig.java
@@ -1,0 +1,23 @@
+package kvansipto.telegram.microservice.config;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@Getter
+@Setter
+@Valid
+@ConfigurationProperties("telegram.bot")
+public class BotConfig {
+
+  @NotBlank
+  String botName;
+  @NotBlank
+  String botToken;
+  @NotBlank
+  Long ownerChatId;
+}
