+++ telegram/src/main/java/kvansipto/telegram/microservice/services/command/ExerciseCommand.java
