+++ telegram/src/main/java/kvansipto/telegram/microservice/services/command/menu/StartCommand.java
