+++ telegram/src/main/java/kvansipto/telegram/microservice/services/wrapper/BotApiMethodInterface.java
