+++ telegram/src/main/java/kvansipto/telegram/microservice/services/wrapper/SendPhotoWrapper.java
